Fixed flag submission functionality
This is going to be a large commit, but it contains mostly little
fixes. The biggest ones include switching flag submission to the
WTForms module, and creating a submissions table, named
CleverName. We still can't create challenges unless we modify the
db directly, but we can now submit flags to challenges created
that way.